requirements.txt
annotated-types==0.7.0

anyio==4.9.0
click==8.1.8
dnspython==2.7.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
motor==3.7.0
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
requests==2.31.0
